requirements.txt
black==24.1.1
flake8==7.1.0
isort==5.13.2
ipykernel~=6.29.2
ipywidgets==8.1.3
jupyterlab~=4.1.6
mypy==1.8.0
numpy==1.26.4
omegaconf==2.3.0
polars~=0.20.26
pylint==3.0.3
pytest==7.4.4
python-dotenv==1.0.1
pytube==15.0.0
requests~=2.31.0
scikit-learn==1.2.2
sentence-transformers==2.6.1
torch
tqdm==4.66.4
youtube-transcript-api==0.6.2